Added support for Flex and Bison
We have to fight qmake because if we use the normal rules for yacc and lex the
results will not compile, due to file renaming.

The new rules contained in the .prf files are ok, and will generate
'FILE.lex.cpp' from 'FILE.l' and 'FILE.tab.hpp' and 'FILE.tab.cpp' from
'FILE.ypp'.

We also set git to ignore the autogenerated files by flex and bison.